+++ pangea-sdk/service/ip_intel/api.go
@@ -188,6 +188,9 @@ func (e *ipIntel) IsProxy(ctx context.Context, input *IpProxyRequest) (*pangea.P
 }
 
 type IpGeolocateRequest struct {
+	// Base request has ConfigID for multi-config projects
+	pangea.BaseRequest
+
 	Ip       string `json:"ip"`
 	Verbose  *bool  `json:"verbose,omitempty"`
 	Raw      *bool  `json:"raw,omitempty"`
@@ -195,6 +198,9 @@ type IpGeolocateRequest struct {
 }
 
 type IpReputationRequest struct {
+	// Base request has ConfigID for multi-config projects
+	pangea.BaseRequest
+
 	Ip       string `json:"ip"`
 	Verbose  *bool  `json:"verbose,omitempty"`
 	Raw      *bool  `json:"raw,omitempty"`
@@ -229,6 +235,9 @@ type GeolocateData struct {
 }
 
 type IpDomainRequest struct {
+	// Base request has ConfigID for multi-config projects
+	pangea.BaseRequest
+
 	Ip       string `json:"ip"`
 	Verbose  *bool  `json:"verbose,omitempty"`
 	Raw      *bool  `json:"raw,omitempty"`
@@ -247,6 +256,9 @@ type DomainData struct {
 }
 
 type IpVPNRequest struct {
+	// Base request has ConfigID for multi-config projects
+	pangea.BaseRequest
+
 	Ip       string `json:"ip"`
 	Verbose  *bool  `json:"verbose,omitempty"`
 	Raw      *bool  `json:"raw,omitempty"`
@@ -264,6 +276,9 @@ type VPNData struct {
 }
 
 type IpProxyRequest struct {
+	// Base request has ConfigID for multi-config projects
+	pangea.BaseRequest
+
 	Ip       string `json:"ip"`
 	Verbose  *bool  `json:"verbose,omitempty"`
 	Raw      *bool  `json:"raw,omitempty"`