+++ CHANGELOG.md
@@ -5,6 +5,33 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.0] - Unreleased
+
+# Added
+
+# Changed
+
+- Audit.Log() now receive a IEvent interface instead of Event to log
+- Audit.Search() and .Result() receive an additional param IEvent to unmarshal events
+- Make service structs private to force using its interface
+- Rename Request/Result structs to user Request/Result postfix instead of Input/Output
+- Update multiple fields to do not be pointers unnecesarily
+- Vault field RotationState is now a ItemVersionState
+
+# Deleted
+
+- All Intel deprecated methods
+
+
+
+## [1.10.0] - 2023-06-26
+
+# Added
+- Multiconfig support
+- Instructions to setup token and domain in examples
+
+
+## [1.9.1] - 2023-06-09
 # Added
 
 - Defang examples
@@ -136,7 +163,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Redact client
 
 
-[unreleased]: https://github.com/pangeacyber/pangea-go/compare/pangea-sdk/v1.9.1...main
+[unreleased]: https://github.com/pangeacyber/pangea-go/compare/pangea-sdk/v1.10.0...main
+[1.10.0]: https://github.com/pangeacyber/pangea-go/compare/pangea-sdk/v1.9.1...v1.10.0
 [1.9.1]: https://github.com/pangeacyber/pangea-go/compare/pangea-sdk/v1.9.0...v1.9.1
 [1.9.0]: https://github.com/pangeacyber/pangea-go/compare/pangea-sdk/v1.8.0...v1.9.0
 [1.8.0]: https://github.com/pangeacyber/pangea-go/compare/pangea-sdk/v1.7.0...v1.8.0